Change the drawing component page navigation methods
This change makes all the navigation methods only take $self as an
argument rather than having `$button` as the first argument as expected in
a callback. This makes these functions usable outside of a callback
since they do not use the `$button` object.

This also means that all the button callbacks need to be replaced with
an anonymous coderef.