Outline: Change page number for outline to be lax
This introduces a `LaxPageNumber` type that can be a positive or zero
integer. This allows for data that may point to an invalid page.

We must therefore check for whether we can go to that page when using
the GUI which is why `on_tree_view_row_activate_cb` introduces a check
for valid `PageNumber` data.