Create a ::Page::PDF model
The `::Page::PDF` replaces the use of the `::Page::RenderedFromPNG` that
was previously used by `::Document::PDF`. The purpose of this
replacement is that the bounds calculation no longer depends upon
rendering the data on to a `Cairo::ImageSurface` first. Instead, the
rendering is done lazily, so there should be no performance degradation.

Another benefit of this approach is that the substitution will work
without having to change any of the other code.